Browse btn corrections
Browse butten has had a spelling correction as well as a fix to an issue where the build did not correctly set its visable to true. I have decided now to instead change the browse butten to be set to inactive when not used instead of setting it to not visable.