Icon added as well as testing a light bug fixs
An icon file has been added so that there is an icon when the program has been launched.
Light testing completed with latest patch work. I have also corrected some minor issues such as the display of default instructional text in any text box that is being used if the user changes the setting.
The clear butten will now also clear the IV and Key text boxs when they are active. If they are not active though they are left with the default text. This default text discribes what the feild requires to reduce confustion with the user.