Multiple files can now be encrypted together
Added an option before selecting a file to encrypt/decrypt to choose multiple files at once. The code will still perform the encryption one at a time but it will populate a list with the paths the user has chosen. They are then cycled through using a foreach loop to run the encryption/decryption on each path. If the user wishes to encrypt multiple files using different keys or IVs they will have to do them individually as the program will use the same key and IV for all selected files using the multi select option.